requirements.txt
GitPython~=3.1.44
pandas~=2.0.3
numpy~=1.24.4
protobuf~=5.29.3
google-api-python-client~=2.160.0